Amend round_to_power_two to support 64 bit integers.
Also change from the bit-twiddling version to using the builtins to
count leading zeroes.

+++ klib/misc/math_hacks.h
@@ -1,6 +1,6 @@
 #ifndef KLIB_MATH_HACKS
 #define KLIB_MATH_HACKS
 
-unsigned int round_to_power_two(unsigned int input);
+unsigned long round_to_power_two(unsigned long input);
 
 #endif
